endian: clean up endian code, use builtin.h and exact-int.h
This is a massive simplification of the API, reducing everything to six
functions.  Dropping the bswap and exact-width integer stuff helped,
but I also dropped a bunch of macros which were basically just
duplicates of these six functions.

It's still possible to use endian.h without the other headers, you just
need to provide unsigned 16/32/64 bit types and bswap implementations.