builtin: add choose_expr/types_compatible_p-based suffix detection
The existing suffix detection code will still be used as a fallback for
compilers lacking the __builtin_choose_expr and
__builtin_types_compatible_p builtins, but on compilers which do have
those builtins this code should help us avoid some warnings when the C
compiler thinks types are incompatible even though they have the same
signedness and min/max values.

For example, on OS X int64_t is compatible with long long but not long,
even though sizeof(long) == sizeof(long long) (and both are 64 bits).
Simply reversing the check direction to prefer long long over long for
64 bit types resolves the issue on OS X but causes warnings on Windows
instead.

If this is insufficient, my only remaining idea is to hardcode
the PSNIP_BUILTIN_VARIANT*_* on some platforms.  This isn't a big deal,
and I'm willing to add the necessary code, but it will require feedback
from people working on those platforms.