builtin: add some casts to silence an Apple LLVM warning
Apple LLVM doesn't like it when you pass a long* to a long long*, even
when long and long long are compatible.  This is patch just adds a
cast to void* in the exact-width wrappers of addc and subc.
Unfortunately this means you could also be passing int*, or anything
else, and the compiler won't complain.

The real solution is to find a more reliable way to determine which
suffix to use; i.e., if int64_t is a long long we should use
__builtin_*ll instead of __builtin_*l (or vice versa), even when the
types in question are compatible.

The only other thing I can come up with is to generate wrapper
functions for the exact size versions:

  T1 psnip_builtin_addc64(T1 x, T1 y, T1 ci, T1* co) {
    T2 tmp;
    T1 ret = (T1) __builtin_addcX((T2) x, (T2) y, (T2) ci, &tmp);
    *co = (T1) tmp;
    return ret;
  }

I'm not sure how good compilers are about optimizing this, though.
Since this is potentially performance-sensitive code, I don't want to
do it without looking very carefully about what compilers do.