builtin: use psnip_uint32_t for 32-bit intrinsics
Windows is LLP64, and uses long as a 32-bit type.  However, most other
platforms are LP64, and this assumption doesn't hold.  This patch
changes the 32-bit intrinsics to use exact-width types.

Unfortunately, this triggers some warnings (C4057) for functions which
take a long*.  For psnip_intrin_* versions we can take care of that
by disabling that warning for each call, but if people use the
intrinsic directly (i.e., _bittest instead of psnip_intrin_bittest)
there isn't much we can do.