tests/clock: allow a bit more slop in accuracy tests
A FreeBSD guest (qemu-kvm) on my desktop has trouble passing the old
version.  I'm not sure if the sleep call or the clock is the problem,
but the test is really just meant as a sanity check.  There isn't
much we can do about sleep taking too long, or a clock not being as
accurate as we'd like.