clock: don't try to use CLOCK_MONOTONIC_RAW or _PRECISE
Just use CLOCK_MONOTONIC. It can be significantly faster, and is
generally accurate enough for our purposes.

If anyone has a problem with this, I'd be willing to add the ability
to define something like PSNIP_CLOCK_PREFER_PRECISE or
PSNIP_CLOCK_PREFER_FAST to have the code prefer either fast or
precise clocks instead of the defaults.